requirements.txt
Flask>=3.1,<4
google-api-python-client>=2.166,<3
google-auth-httplib2>=0.2.0,<1
google-auth-oauthlib>=1.2,<2
reportlab>=4.0,<5
Pillow>=10.0,<12
